requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements-dev.in
#
amqp==5.2.0               # via -r requirements.txt, kombu
astroid==2.14.2           # via pylint
attrs==20.3.0             # via flake8-bugbear
backoff==2.2.1            # via -r requirements.txt, opentelemetry-exporter-otlp-proto-common, opentelemetry-exporter-otlp-proto-grpc, opentelemetry-exporter-otlp-proto-http
billiard==4.2.0           # via -r requirements.txt, celery
black==24.3.0             # via -r requirements-dev.in
blinker==1.4              # via -r requirements.txt, elastic-apm, sentry-sdk
boto3==1.17.97            # via -r requirements.txt
botocore==1.20.97         # via -r requirements.txt, boto3, s3transfer
celery==5.4.0             # via -r requirements.txt, dbt-copilot-python
certifi==2024.8.30        # via -r requirements.txt, elastic-apm, requests, sentry-sdk
charset-normalizer==3.1.0  # via -r requirements.txt, requests
click-didyoumean==0.3.1   # via -r requirements.txt, celery
click-plugins==1.1.1      # via -r requirements.txt, celery
click-repl==0.3.0         # via -r requirements.txt, celery
click==8.1.3              # via -r requirements.txt, black, celery, click-didyoumean, click-plugins, click-repl, flask, pip-tools
dbt-copilot-python==0.2.1  # via -r requirements.txt
deprecated==1.2.14        # via -r requirements.txt, opentelemetry-api, opentelemetry-exporter-otlp-proto-grpc, opentelemetry-exporter-otlp-proto-http
dill==0.3.6               # via pylint
ecs-logging==0.5.0        # via -r requirements.txt
elastic-apm[flask]==5.10.0  # via -r requirements.txt
exceptiongroup==1.2.1     # via pytest
flake8-bugbear==23.2.13   # via -r requirements-dev.in
flake8==6.0.0             # via -r requirements-dev.in, flake8-bugbear
flask==2.2.5              # via -r requirements.txt, sentry-sdk
gevent==23.9.1            # via -r requirements.txt
googleapis-common-protos==1.63.0  # via -r requirements.txt, opentelemetry-exporter-otlp-proto-grpc, opentelemetry-exporter-otlp-proto-http
greenlet==3.0.3           # via -r requirements.txt, gevent, playwright
grpcio==1.62.2            # via -r requirements.txt, opentelemetry-exporter-otlp-proto-grpc
idna==3.7                 # via -r requirements.txt, requests
importlib-metadata==6.8.0  # via -r requirements.txt, flask, opentelemetry-api
iniconfig==2.0.0          # via pytest
ipy==0.83                 # via -r requirements.txt
isort==4.3.21             # via pylint
itsdangerous==2.1.2       # via -r requirements.txt, flask
jinja2==3.1.4             # via -r requirements.txt, flask
jmespath==0.10.0          # via -r requirements.txt, boto3, botocore
kombu==5.3.7              # via -r requirements.txt, celery
lazy-object-proxy==1.4.3  # via astroid
lxml==4.9.1               # via -r requirements.txt
markupsafe==2.1.2         # via -r requirements.txt, jinja2, sentry-sdk, werkzeug
mccabe==0.7.0             # via flake8, pylint
mypy-extensions==1.0.0    # via black
opentelemetry-api==1.22.0  # via -r requirements.txt, opentelemetry-distro, opentelemetry-exporter-otlp-proto-grpc, opentelemetry-exporter-otlp-proto-http, opentelemetry-instrumentation, opentelemetry-instrumentation-wsgi, opentelemetry-propagator-aws-xray, opentelemetry-sdk
opentelemetry-distro==0.43b0  # via -r requirements.txt, dbt-copilot-python
opentelemetry-exporter-otlp-proto-common==1.22.0  # via -r requirements.txt, opentelemetry-exporter-otlp-proto-grpc, opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.22.0  # via -r requirements.txt, opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.22.0  # via -r requirements.txt, opentelemetry-exporter-otlp
opentelemetry-exporter-otlp==1.22.0  # via -r requirements.txt, dbt-copilot-python
opentelemetry-instrumentation-wsgi==0.43b0  # via -r requirements.txt, dbt-copilot-python
opentelemetry-instrumentation==0.43b0  # via -r requirements.txt, opentelemetry-distro, opentelemetry-instrumentation-wsgi
opentelemetry-propagator-aws-xray==1.0.1  # via -r requirements.txt, dbt-copilot-python
opentelemetry-proto==1.22.0  # via -r requirements.txt, opentelemetry-exporter-otlp-proto-common, opentelemetry-exporter-otlp-proto-grpc, opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk-extension-aws==2.0.1  # via -r requirements.txt, dbt-copilot-python
opentelemetry-sdk==1.22.0  # via -r requirements.txt, opentelemetry-distro, opentelemetry-exporter-otlp-proto-grpc, opentelemetry-exporter-otlp-proto-http, opentelemetry-sdk-extension-aws
opentelemetry-semantic-conventions==0.43b0  # via -r requirements.txt, opentelemetry-instrumentation-wsgi, opentelemetry-sdk
opentelemetry-util-http==0.43b0  # via -r requirements.txt, opentelemetry-instrumentation-wsgi
packaging==23.1           # via black, pytest
pathspec==0.11.0          # via black
pip-tools==5.4.0          # via -r requirements-dev.in
pipdeptree==2.15.1        # via -r requirements-dev.in
platformdirs==3.0.0       # via black, pylint
playwright==1.43.0        # via pytest-playwright
pluggy==1.5.0             # via pytest
prompt-toolkit==3.0.43    # via -r requirements.txt, click-repl
protobuf==4.25.3          # via -r requirements.txt, googleapis-common-protos, opentelemetry-proto
pycodestyle==2.10.0       # via flake8
pyee==11.1.0              # via playwright
pyflakes==3.0.1           # via flake8
pylint==2.16.3            # via -r requirements-dev.in
pytest-base-url==2.1.0    # via pytest-playwright
pytest-playwright==0.4.4  # via -r requirements-dev.in
pytest==8.2.0             # via pytest-base-url, pytest-playwright
python-dateutil==2.9.0.post0  # via -r requirements.txt, botocore, celery
python-dotenv==1.0.1      # via -r requirements-dev.in
python-slugify==8.0.4     # via pytest-playwright
requests==2.32.0          # via -r requirements.txt, dbt-copilot-python, opentelemetry-exporter-otlp-proto-http, pytest-base-url
s3transfer==0.4.2         # via -r requirements.txt, boto3
sentry-sdk[flask]==2.13.0  # via -r requirements.txt
six==1.15.0               # via -r requirements.txt, pip-tools, python-dateutil
text-unidecode==1.3       # via python-slugify
tomli==2.0.1              # via black, pylint, pytest
tomlkit==0.11.6           # via pylint
typing-extensions==4.11.0  # via -r requirements.txt, astroid, black, kombu, opentelemetry-sdk, pylint
tzdata==2024.1            # via -r requirements.txt, celery
urllib3==1.26.19          # via -r requirements.txt, botocore, elastic-apm, requests, sentry-sdk
vine==5.1.0               # via -r requirements.txt, amqp, celery, kombu
wcwidth==0.2.13           # via -r requirements.txt, prompt-toolkit
werkzeug==3.0.6           # via -r requirements.txt, flask
wrapt==1.16.0             # via -r requirements.txt, astroid, deprecated, opentelemetry-instrumentation
zipp==3.19.1              # via -r requirements.txt, importlib-metadata
zope.event==5.0           # via -r requirements.txt, gevent
zope.interface==6.1       # via -r requirements.txt, gevent

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
